Add flight filtering
...and handle query params a bit nicer.

We've stopped `replace`ing the default query params into the URL. That got
really broken at some point. So we just use defaults locally, when nothing it
set and it works fine. And we don't stomp on the end-user's URL.

We have set a default trophy in the trophies description object. This is used if
a trophy was not specified.

We've also set a season start date in the trophies object. This allows us to
request flights for a given date range, and the API sorts out how many requests
to make to the BGA Ladder. This works really nicely, but still doesn't solve the
Kellman clock situation where the date range is different from the general case.
I think the solution is to allow a trophy to specify its own date range.
Initially I was going to set fields in the trophy object, but maybe the trophy
expression can already do this?

Finally, we added a link to the BGA Ladder site for each result. The styling is
wrong though.